Generate apps from specs (bug 1065616)
