+++ data/apps/test_apps.json
@@ -1,0 +1,35 @@
+[{"type": "hosted", "status": "public"},
+ {"type": "hosted", "status": "pending"},
+ {"type": "hosted", "status": "waiting"},
+ {"type": "hosted", "status": "incomplete"},
+ {"type": "hosted", "status": "unlisted"},
+ {"type": "hosted", "status": "deleted"},
+ {"type": "hosted", "status": "incomplete"},
+ {"type": "hosted", "status": "public", "num_ratings": 0},
+ {"type": "hosted", "status": "public", "num_ratings": 100},
+ {"type": "hosted", "status": "public", "num_previews": 5},
+ {"type": "hosted", "status": "public", "num_locales": 5},
+ {"type": "packaged", "status": "public"},
+ {"type": "packaged", "status": "pending"},
+ {"type": "packaged", "status": "waiting"},
+ {"type": "packaged", "status": "incomplete"},
+ {"type": "packaged", "status": "unlisted"},
+ {"type": "packaged", "status": "deleted"},
+ {"type": "packaged", "status": "incomplete"},
+ {"type": "packaged", "status": "public", "num_ratings": 0},
+ {"type": "packaged", "status": "public", "num_ratings": 100},
+ {"type": "packaged", "status": "public", "num_previews": 5},
+ {"type": "packaged", "status": "public", "num_locales": 5},
+ {"type": "packaged", "status": "public", "versions": ["public", "public", "public"]},
+ {"type": "packaged", "status": "public", "versions": ["public", "pending"]},
+ {"type": "packaged", "status": "public", "versions": ["public", "disabled"]},
+ {"type": "packaged", "status": "public", "versions": ["public", "deleted"]},
+ {"type": "packaged", "status": "public", "versions": ["public", "waiting"]},
+ {"type": "packaged", "status": "incomplete", "versions": ["deleted", "deleted"]},
+ {"type": "privileged", "status": "public", "permissions": ["camera", "storage"]},
+ {"type": "privileged", "status": "pending", "permissions": ["camera", "storage"]},
+ {"type": "privileged", "status": "waiting", "permissions": ["camera", "storage"]},
+ {"type": "privileged", "status": "incomplete", "permissions": ["camera", "storage"]},
+ {"type": "privileged", "status": "unlisted", "permissions": ["camera", "storage"]},
+ {"type": "privileged", "status": "deleted", "permissions": ["camera", "storage"]},
+ {"type": "privileged", "status": "incomplete", "permissions": ["camera", "storage"]}]
