+++ docs/topics/install-zamboni/installation.rst
@@ -158,10 +158,17 @@ migrations like this::
 
 More info on schematic: https://github.com/mozilla/schematic
 
-Creating Test Apps
+Loading Test Apps
 ~~~~~~~~~~~~~~~~~~
 
-Example applications for testing purposes can be generated by running::
+Example applications for testing purposes can be loaded by running::
+
+    ./manage.py generate_apps_from_spec data/apps/test_apps.json
+
+See :doc:`/topics/fake-app-spec` for details of the JSON format.
+
+If you just want a certain number of public apps in various categories to be
+created, run::
 
     ./manage.py generate_apps N
