+++ mkt/webapps/tasks.py
@@ -1,15 +1,10 @@
 import datetime
 import hashlib
-import itertools
 import json
 import logging
 import os
-import random
 import shutil
-import StringIO
 import subprocess
-import tempfile
-from zipfile import ZipFile, ZIP_DEFLATED
 
 from django.conf import settings
 from django.contrib.auth.models import AnonymousUser
@@ -23,26 +18,23 @@ import requests
 from celery import chord
 from celery.exceptions import RetryTaskError
 from celeryutils import task
-from PIL import Image
 from requests.exceptions import RequestException
 from tower import ugettext as _
 
 import mkt
 from lib.post_request_task.task import task as post_request_task
 from mkt.abuse.models import AbuseReport
-from mkt.constants.categories import CATEGORY_CHOICES
 from mkt.constants.regions import RESTOFWORLD
 from mkt.developers.models import ActivityLog, AppLog
 from mkt.developers.tasks import (_fetch_manifest, fetch_icon, pngcrush_image,
-                                  resize_preview, save_icon, validator)
+                                  resize_preview, validator)
 from mkt.files.models import FileUpload
 from mkt.files.utils import WebAppParser
-from mkt.ratings.models import Review
 from mkt.reviewers.models import EscalationQueue, RereviewQueue
 from mkt.site.decorators import set_task_user, use_master, write
 from mkt.site.helpers import absolutify
 from mkt.site.mail import send_mail_jinja
-from mkt.site.utils import chunked, JSONEncoder, slugify
+from mkt.site.utils import chunked, JSONEncoder
 from mkt.users.models import UserProfile
 from mkt.users.utils import get_task_user
 from mkt.webapps.indexers import WebappIndexer
@@ -673,214 +665,6 @@ def set_storefront_data(app_id, disable=False, **kw):
         return
 
     app.set_iarc_storefront_data(disable=disable)
-
-
-adjectives = ['Exquisite', 'Delicious', 'Elegant', 'Swanky', 'Spicy',
-              'Food Truck', 'Artisanal', 'Tasty']
-nouns = ['Sandwich', 'Pizza', 'Curry', 'Pierogi', 'Sushi', 'Salad', 'Stew',
-         'Pasta', 'Barbeque', 'Bacon', 'Pancake', 'Waffle', 'Chocolate',
-         'Gyro', 'Cookie', 'Burrito', 'Pie']
-fake_app_names = list(itertools.product(adjectives, nouns))[:-1]
-
-
-def generate_app_data(num):
-    repeats, tailsize = divmod(num, len(fake_app_names))
-    if repeats:
-        apps = fake_app_names[:]
-        for i in range(repeats - 1):
-            for a in fake_app_names:
-                apps.append(a + (str(i + 1),))
-        for a in fake_app_names[:tailsize]:
-            apps.append(a + (str(i + 2),))
-    else:
-        apps = random.sample(fake_app_names, tailsize)
-    # Let's have at least 3 apps in each category, if we can.
-    if num < (len(CATEGORY_CHOICES) * 3):
-        num_cats = max(num // 3, 1)
-    else:
-        num_cats = len(CATEGORY_CHOICES)
-    catsize = num // num_cats
-    ia = iter(apps)
-    for cat_slug, cat_name in CATEGORY_CHOICES[:num_cats]:
-        for n in range(catsize):
-            appname = ' '.join(next(ia))
-            yield (appname, cat_slug)
-    for i, app in enumerate(ia):
-        appname = ' '.join(app)
-        cat_slug, cat_name = CATEGORY_CHOICES[i % len(CATEGORY_CHOICES)]
-        yield (appname, cat_slug)
-
-
-def generate_icon(app):
-    im = Image.new(
-        "RGB", (128, 128),
-        "#" + hashlib.md5(unicode(app.name).encode('utf8')).hexdigest()[:6])
-    f = StringIO.StringIO()
-    im.save(f, 'png')
-    save_icon(app, f.getvalue())
-
-
-def generate_preview(app, n=1):
-    im = Image.new(
-        "RGB", (320, 480),
-        "#" + hashlib.md5(
-            unicode(app.name).encode('utf8') + chr(n)).hexdigest()[:6])
-    p = Preview.objects.create(addon=app, filetype="image/png",
-                               thumbtype="image/png",
-                               caption="screenshot " + str(n),
-                               position=n)
-    f = tempfile.NamedTemporaryFile()
-    im.save(f, 'png')
-    resize_preview(f.name, p)
-
-
-def generate_translations(app):
-    fr_prefix = u'(fran\xe7ais) '
-    es_prefix = u'(espa\xf1ol) '
-    oldname = unicode(app.name)
-    app.name = {'en': oldname,
-                'fr': fr_prefix + oldname,
-                'es': es_prefix + oldname}
-    app.save()
-
-
-def generate_ratings(app, num):
-    for n in range(num):
-        email = 'testuser%s@example.com' % (n,)
-        user, _ = UserProfile.objects.get_or_create(
-            username=email, email=email, source=mkt.LOGIN_SOURCE_UNKNOWN,
-            display_name=email)
-        Review.objects.create(
-            addon=app, user=user, rating=random.randrange(0, 6),
-            title="Test Review " + str(n), body="review text")
-
-
-def generate_hosted_app(name, category):
-    # Let's not make production code depend on stuff in the test package --
-    # importing it only when called in local dev is fine.
-    from mkt.site.tests import app_factory
-    a = app_factory(categories=[category], name=name, complete=True,
-                    rated=True)
-    a.versions.latest().update(reviewed=datetime.datetime.now())
-    generate_hosted_manifest(a)
-    return a
-
-
-def generate_hosted_manifest(app):
-    data = {
-        'name': unicode(app.name),
-        'description': 'This app has been automatically generated',
-        'version': '1.0',
-        'icons': {
-            '16': 'http://testmanifest.com/icon-16.png',
-            '48': 'http://testmanifest.com/icon-48.png',
-            '128': 'http://testmanifest.com/icon-128.png'
-        },
-        'installs_allowed_from': ['*'],
-        'developer': {
-            'name': 'Marketplace Team',
-            'url': 'https://marketplace.firefox.com/credits'
-        }
-    }
-    AppManifest.objects.create(
-        version=app.latest_version, manifest=json.dumps(data))
-    app.update(manifest_url='http://%s.testmanifest.com/manifest.webapp' %
-               (slugify(unicode(app.name)),))
-
-
-def generate_app_package(app, out, permissions, version='1.0'):
-    fr_prefix = u'(fran\xe7ais) '
-    es_prefix = u'(espa\xf1ol) '
-    manifest = {
-        'version': version.version,
-        'name': unicode(app.name),
-        'description': ('This packaged app has been automatically generated'
-                        ' (version %s)' % (version.version,)),
-        'icons': {
-            '16': '/icons/16.png',
-            '32': '/icons/32.png',
-            '256': '/icons/256.png'
-        },
-        'developer': {
-            'name': 'Marketplace Team',
-            'url': 'https://marketplace.firefox.com/credits'
-        },
-        'installs_allowed_launch': ['*'],
-        'from_path': 'index.html',
-        'locales': {
-            'es': {
-                'name': es_prefix + unicode(app.name),
-                'description': 'This packaged app has been automatically'
-                               ' generated'
-            },
-            'fr': {
-                'name': fr_prefix + unicode(app.name),
-                'description': 'This packaged app has been automatically'
-                               ' generated'
-            },
-        },
-        'permissions': dict(((k, {"description": k})
-                             for k in permissions)),
-        'default_locale': 'en',
-        'orientation': 'landscape',
-        'type': 'privileged' if permissions else 'web',
-        'fullscreen': 'true'
-    }
-    with ZipFile(file=out, mode='w', compression=ZIP_DEFLATED) as outz:
-        for size in ('32', 'med'):
-            outz.writestr(
-                'icons/%s.png' % (size,),
-                open(os.path.join(
-                    settings.MEDIA_ROOT,
-                    'img/app-icons/%s/generic.png' % (size,))).read())
-        outz.writestr('script.js',
-                      'document.onload=function() {alert("Hello!");};')
-        outz.writestr(
-            'index.html',
-            '<title>Packaged app</title><script src="script.js"></script>'
-            '<h1>Test packaged app</h1>')
-        outz.writestr("manifest.webapp", json.dumps(manifest))
-    AppManifest.objects.create(
-        version=version, manifest=json.dumps(manifest))
-
-
-def generate_packaged_app(name, category, permissions=(), num_versions=1):
-    from mkt.site.tests import app_factory, version_factory
-    app = app_factory(categories=[category], name=name, complete=True,
-                      rated=True, is_packaged=True,
-                      version_kw={'version': '1.0'})
-    f = app.latest_version.all_files[0]
-    f.update(filename=f.generate_filename())
-    fp = os.path.join(app.latest_version.path_prefix, f.filename)
-    try:
-        os.makedirs(os.path.dirname(fp))
-    except OSError:
-        pass
-    with open(fp, 'w') as out:
-        generate_app_package(app, out, permissions=permissions,
-                             version=app.latest_version)
-        for i in range(1, num_versions):
-            v = version_factory(version="1." + str(i), addon=app)
-            generate_app_package(app, out, permissions, v)
-    return app
-
-
-def generate_apps(hosted=0, packaged=0, privileged=0, versions=1):
-    apps = generate_app_data(hosted + packaged + privileged)
-    for i, (appname, cat_slug) in enumerate(apps):
-        if i < privileged:
-            app = generate_packaged_app(appname, cat_slug,
-                                        num_versions=versions,
-                                        permissions=['camera', 'storage'])
-        elif i < (privileged + packaged):
-            app = generate_packaged_app(appname, cat_slug,
-                                        num_versions=versions)
-        else:
-            app = generate_hosted_app(appname, cat_slug)
-        generate_icon(app)
-        generate_preview(app)
-        generate_translations(app)
-        generate_ratings(app, 5)
 
 
 @task
